requirements.txt
flask
flask-markdown
flask-mysql
markdown
flask_sqlalchemy
pillow